refactor xud dir creation for nested path creation
