+++ lib/Config.ts
@@ -2,7 +2,7 @@ import fs from 'fs';
 import os from 'os';
 import path from 'path';
 import toml from 'toml';
-import { deepMerge } from './utils/utils';
+import { deepMerge, mkdirRecursiveSync } from './utils/utils';
 import { PoolConfig } from './p2p/Pool';
 import { LndClientConfig } from './lndclient/LndClient';
 import { RaidenClientConfig } from './raidenclient/RaidenClient';
@@ -88,7 +88,7 @@ class Config {
 
   public async load() {
     if (!fs.existsSync(this.xudir)) {
-      fs.mkdirSync(this.xudir);
+      mkdirRecursiveSync(this.xudir);
     } else if (fs.existsSync(`${this.xudir}xud.conf`)) {
       const configText = fs.readFileSync(`${this.xudir}xud.conf`, 'utf8');
       try {
